+++ src/components/Footer/Footer.jsx
@@ -15,19 +15,29 @@ class Footer extends React.Component {
             <nav className="footer-nav">
               <ul>
                 <li>
-                  <a href="https://www.creative-tim.com" target="_blank">Creative Tim</a>
+                  <a href="https://www.creative-tim.com" target="_blank">
+                    Creative Tim
+                  </a>
                 </li>
                 <li>
-                  <a href="https://blog.creative-tim.com" target="_blank">Blog</a>
+                  <a href="https://blog.creative-tim.com" target="_blank">
+                    Blog
+                  </a>
                 </li>
                 <li>
-                  <a href="https://www.creative-tim.com/license" target="_blank">Licenses</a>
+                  <a
+                    href="https://www.creative-tim.com/license"
+                    target="_blank"
+                  >
+                    Licenses
+                  </a>
                 </li>
               </ul>
             </nav>
             <div className="credits ml-auto">
               <div className="copyright">
-                &copy; {1900 + new Date().getYear()}, made with <i className="fa fa-heart heart"></i> by Creative Tim
+                &copy; {1900 + new Date().getYear()}, made with{" "}
+                <i className="fa fa-heart heart" /> by Creative Tim
               </div>
             </div>
           </Row>