+++ src/index.js
@@ -13,11 +13,12 @@ const hist = createBrowserHistory();
 
 ReactDOM.render(
   <React.StrictMode>
-  <Router history={hist}>
-    <Switch>
-      <Route path="/admin" render={(props) => <AdminLayout {...props} /> } />
-      <Redirect to="/admin/dashboard" />
-    </Switch>
-  </Router></React.StrictMode>,
+    <Router history={hist}>
+      <Switch>
+        <Route path="/admin" render={props => <AdminLayout {...props} />} />
+        <Redirect to="/admin/dashboard" />
+      </Switch>
+    </Router>
+  </React.StrictMode>,
   document.getElementById("root")
 );